+++ client/components/PageLayout.js
@@ -1,14 +1,10 @@
-import { Card, Layout, Page } from "@shopify/polaris";
+import { Layout, Page } from "@shopify/polaris";
 import React from "react";
 
 function PageLayout({ children }) {
   return (
     <Page>
-      <Layout>
-        <Layout.Section>
-          <Card>{children}</Card>
-        </Layout.Section>
-      </Layout>
+      <Layout>{children}</Layout>
     </Page>
   );
 }
