+++ client/index.js
@@ -1,5 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import MyApp from "./App";
+import App from "./components/App";
 
-ReactDOM.render(<MyApp />, document.getElementById("root"));
+ReactDOM.render(<App />, document.getElementById("root"));
